// packages/contracts/src/enums.ts
export enum LeadStatus {
  NEW = "new",
  FOLLOWING = "following",
  QUOTED = "quoted",
  CONVERTED = "converted",
  INVALID = "invalid",
}

export enum OrderStatus {
  PENDING_CONFIRM = "pending_confirm",
  PENDING_ASSIGN = "pending_assign",
  ASSIGNED = "assigned",
  SERVING = "serving",
  COMPLETED = "completed",
  CANCELLED = "cancelled",
  AFTER_SALE = "after_sale",
}